+++ 26.2/src/main/java/com/minecolonies/core/event/EventHandler.java
@@ -452,7 +452,7 @@ public class EventHandler
         {
             for (final IColony colony : IColonyManager.getInstance().getAllColonies())
             {
-                if (colony.getPermissions().getRank(player).isColonyManager())
+                if (colony.getWorld() != null && colony.getPermissions().getRank(player).isColonyManager())
                 {
                     colony.getPackageManager().addImportantColonyPlayer(player);
                     colony.getPackageManager().sendColonyViewPackets();